fix: address Claude review — 5 P1 bugs + 2 P2 issues
P1-1: Fix WGSL type mismatch in opacity.f16.wgsl
  color.a (f32) * f16(u.opacity) → f16(color.a) * f16(u.opacity)
  Would fail validation on all shader-f16 devices.

P1-2: Fix storage.b aliasing after source normalization
  encodeSourceNormalize now writes to storage.c instead of storage.b,
  avoiding read+write aliasing with encodeBaseCorrection's first
  ping-pong slot.

P1-5: Fix zebra overlay replacing preview frame instead of compositing
  clipping-overlay.wgsl now takes two inputs in a single pass:
  srcFrame (composited output, passthrough for non-clipped) and
  clipSrc (pre-clamp accumulator, for clip detection). Clipped
  pixels get zebra pattern composited on top of the frame.

P2-7: Fix ColourInspector reactivity bug
  hasMetadata/hasWarnings captured once at component creation.
  Wrapped in createMemo() for reactive updates when clip changes.
  Added createMemo import from solid-js.

Claude also confirmed: opacity aliasing, LUT wiring, torn-write
sequences, pow() clamping, <For> component all already correctly
resolved. Scope dispatch and gamut matrix are documented as deferred
limitations (P1-3, P1-4, P1-6 in Claude's review).

+++ src/engine/shaders/opacity.f16.wgsl
@@ -16,6 +16,6 @@ fn main(@builtin(global_invocation_id) gid: vec3<u32>) {
 
   let coord = vec2<u32>(gid.x, gid.y);
   let color = textureLoad(src, coord, 0);
-  let newAlpha = color.a * f16(u.opacity);
+  let newAlpha = f16(color.a) * f16(u.opacity);
   textureStore(dst, coord, vec4(color.rgb, f32(newAlpha)));
 }
